+++ src/index.js
@@ -30,9 +30,8 @@ function subscribe(
   onError: Error => void,
   onComplete: () => void,
 ): Unsubscribe {
-  // TODO: If this approach works, add markers to Watermelon to indicate Model, Query cleanly
-  if (value.experimentalSubscribe && value._raw) {
-    // HACK: This is a Watermelon Model
+  const wmelonTag = value && value.constructor && value.constructor._wmelonTag
+  if (wmelonTag === 'model') {
     onNext(value)
     return value.experimentalSubscribe(isDeleted => {
       if (isDeleted) {
@@ -41,8 +40,7 @@ function subscribe(
         onNext(value)
       }
     })
-  } else if (value.experimentalSubscribe && value._rawDescription) {
-    // HACK: This is a Watermelon Query
+  } else if (wmelonTag === 'query') {
     return value.experimentalSubscribe(onNext)
   } else if (typeof value.observe === 'function') {
     const subscription = value.observe().subscribe(onNext, onError, onComplete)